error message when deleting item that having relation

+++ app/Http/Controllers/ObatController.php
@@ -61,11 +61,15 @@ class ObatController extends Controller
 
     public function destroyitem($id)
     {
-        Stock::where('item_id', $id)->delete();
         $obatitem = Item::findOrFail($id);
-        // dd($obatuom);
+    
+        // Check if the item is being used in other tables
+        if (Transaction::where('item_id', $id)->count() > 0 || Stock::where('item_id', $id)->count() > 0) {
+            return redirect()->route('wh.iteminv')->with('error', 'You Cant Delete The Item You Trying to Delete!, Because That Item Is under Use for Other Data!');
+        }
+    
         $obatitem->delete();
-
+    
         return redirect()->route('wh.iteminv');
     }
